requirements.txt
fastapi
uvicorn
pydantic
slowapi
requests
python-dotenv
